client_deviceauth: actually implement device status propagation to devauth
missed this TODO during token workflow implementation. without it, no device
can be actually admitted.

also, the correct method to use here is /devices/{id}/status - this is the only
piece of info we're interested in changing.PUT /devices/{id} is in the devauth spec,
but is unnecessary, and is not implemented.

TODO: function names should reflect the 'status update' instead of 'device update'.
leaving this out for now as it's time to finally close this task.

Issues: MEN-411

Signed-off-by: mchalczynski <marcin.chalczynski@open-rnd.pl>