tests/common: fixture for cleaning the deviceauth db
with new tests, we aim for the pattern where each test function
gets a clean db (no dependencies between tests). it's easy in case of
deviceadm db - fixtures for this already exist.

the missing part is cleaning the deviceauth db between tests,
and it's needed for the currently implemented tests - so add
the relevant fixtures.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>